Extensive typo corrections and a style facelift
Solves issue #689 (now closed), with an additional proofreading pass of the entire file for grammar, punctuation, word choices, and spelling.
Clarified some wording related to the Aura throughout.
Clarified the effect and usage of the Focus of Equal Trade, and the Grappler.
Some male-specific language describing the player changed to gender-neutral language (more to do here).
Made sure proper nouns were always capitalized, with nouns designated as proper or not based on their first introduction to the player in documentation.
Replaced mid-sentence hypens and ellipses with the punctuation they implied (more to do here).
Silly change to the description of the Focus Pouch, just for fun.  Fanny packs are so Nineties.  Also clarified how it works.
Changed a few other descriptions.
Removed a misleading line from the Focal Manipulator research, and clarified how it works with a new sentence order.
Spiced up the Node Magnet "warning" section a bit.
Removed misleading text from Arcane stone research, replacing it with something lore-appropriate and related to the way the crafting recipe looks, but not aspect-specific.
Fixed text colors on entries that were missing them.

I intend to do more work on this fork, but a power outage cost me about 3 hours of work.  I'll finish it later.